Structural update and Lucid Toolbox, see extended
Remove included classes in favor of Lucid Toolbox, instead of using the
exact same classes with different names in multiple plugins. Includes
attempt at automatic installation and/or activation of it. Also update
constants and phpdoc, in addition to loading all the plugin parts in the
core class instead of the main file. This has the benefit of only
exposing a single global instance variable, as well as keeping the main
file very clean (which I like for some reason).