Increase specificity and *gasp* add an !important
I feel there will never be a need to have a list style on the slider and
if so, custom CSS can be used instead. I noticed the need for a
specificity increase due to Twenty Twelve's '.entry-content ul' rule
that sets margins and other stuff, overriding the single class name
selectors.